gitool: Initial version
This is a tool for making commits in git.  It displays a list of
modified files and a list of new files (including untracked files).
Users can view the diffs for files by clicking on their names, and
type a commit message into another pane.  Users can check in files,
or undo the checkin, by clicking on the icon next to the filename.
Multiple files can be selected and then checked in with one click.
Files can also be reverted to their previous state using a menu
brought up with the right mouse button.

There is still a lot I want to add to this, but it's already quite
functional.

Signed-off-by: Paul Mackerras <paulus@samba.org>