transport: expose git_tcp_connect() and friends in new tcp.h
Split off a new tcp.c with the functions git_tcp_connect() and
git_proxy_connect() that resolve and connect to a host.

Part of a series to teach git to respect DNS SRV records when making
new connections; this is a preliminary step to make the connection
library easier to understand before adding to it.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>