remove #!interpreter line from shell libraries
In a shell snippet meant to be sourced by another shell script,
an initial #!/bin/sh line is just a comment and has no effect on
the script's functionality.

Some people include such a shebang line anyway, as a familiar way of
documenting what format the file is in.  However, it is even clearer
to include a comment and no shebang at all, to avoid creating the
illusion that the indicated choice of shell will have some effect.
Add some text to git's shell libraries explaining in what context the
fragments are meant to be sourced and remove the shebangs.

As a happy side effect, vim no longer recognizes the file type of
files like git-sh-setup without the ".sh" suffix, which should steer
people toward the git-sh-setup.sh source file.  Thanks to Clemens
Buchacher for noticing.

Complications (thanks to Jeff King for finding these):

- As explained in v1.7.6-rc3~7 (tests: link shell libraries into
  valgrind directory, 2011-06-17), the valgrind support in test
  scripts uses the shebang line to recognize shell libraries which
  should be symlinked into $GIT_VALGRIND/bin/ along with the wrapped
  executables.  This patch tweaks the rule to look for a comment
  starting with "# " in the first line instead.

- Because these files are not marked executable (see v1.7.1-rc0~143^2,
  2010-01-31), the lack of #! line will not confuse the valgrind
  support of our test scripts into replacing the shell libraries with
  the valgrind.sh wrapper (see v1.7.0-rc1~9, 2009-11-25), so this should
  be safe.

Suggested by Russ Allbery through lintian.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>