gitk: pick selection for region blame
Add the new command to the diffmenu, "Show origin of selected region".
The menu command picks selection, and if it exists and covers a single
hunk, request region blame of the corresponding region in the first
parent.

The menu command is disabled if the region blame is impossible, for example
if nothing is selected or the selection does not lie fully in a single diff
hunk.

The decision to always use the first parent is quite voluntary, I just
don't see a real use case in blaming a merge commit. Also, I could not
invent any usable UI for picking a parent.