git: pick selection for region blame
Add the new command to the diffmenu, "Show origin of selected region".
The menu command picks selection, and if it exists and covers a single
hunk, request region blame of the corresponding region in the first
parent.

The menu command could be disabled if the region blame is there is no
valid seection, or the "Show origin of this line" command can be
replaced with it of there is a valid selection. I am not sure which of
these options is better.

The decision to always use the first parent is quite voluntary, I just
don't see a real use case in blaming a merge commit. Also, I could not
invent any usable UI for picking a parent.