gitk: refactor: separated generic hunk parsing out of find_hunk_blamespecs{}
For requesting a region blame, it is necessary to parse a hunk and
find the region in the parent file corresponding to the selected region.
There is already hunk parsin functionality in the find_hunk_blamespec{},
but returns only information for a single line.

The new function, resolve_hunk_lines{}, scans the hunk once and returns
for all hunk lines, in which parent each of them exists awhich number it
has.

This is far more than is required for find_hunk_blamespecs{}, but for
most use cases this should not be a big performance hit.