debian/rules: unset NO_PYTHON=1 setting
The package originally acquired the NO_PYTHON=1 setting in
d81e33ce (debian/rules: do not build Python modules or scripts,
2010-01-28) because it did not have adequate rules to handle the
package with git_remote_helpers helper functions.  Now that
package is gone so this becomes simpler.

We were relying on the NO_PYTHON setting to disable git-p4 and replace
it with a stub that complains.  It's much better not to install git-p4
at all, since that allows people to add such a command in their own
package without git interfering.  Do so.

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>