git-merge-file: do not add LF at EOF while applying unrelated change
If 'current-file' does not contain LF at EOF, and change between
'base-file' and 'other-file' does not change any line close to EOF, the
3-way merge should not add LF to EOF.  This is what 'diff3 -m' does, and
seems to be a reasonable expectation.

The change which introduced the behavior is cd1d61c44f and was about
union merge. The union merges are tested in the t6026-merge-attr, and
merge-file in general are tested in t6023-merge-file. Add tests
"merge conflicting with --.." there, to verify that the fix does not
break anything.

Also add tests "merge without conflict (missing LF at EOF, away from
change in the other file)" and "merge does not add LF away of change",
to demonstrate the changed behavior.